requirements.txt
flask
tensorflow
numpy
pillow
openai